Move from core to common
