+++ common/setup.py
@@ -32,6 +32,7 @@ install_requires = (
         "tqdm",  # version range defined in `core/_setup_utils.py`
         "requests",
         "joblib",  # version range defined in `core/_setup_utils.py`
+        "pyyaml",  # version range defined in `core/_setup_utils.py`
         # s3fs is removed due to doubling install time due to version range resolution
         # "s3fs",  # version range defined in `core/_setup_utils.py`
     ]
